+++ client/src/Components/Nav/MentorUl.js
@@ -1,52 +1,13 @@
-import React, { useState } from "react";
-import { ReactDOM } from "react";
-import Modal from "react-bootstrap/Modal";
-import Button from 'react-bootstrap/Button';
-import Form from 'react-bootstrap/Form';
-
 const MentorUl = () => {
-	const [show, setShow] = useState(false);
-  const [channelAdd, setChannelAdd] = useState('')
-
-	const handleClose = () => setShow(false);
-	const handleShow = () => setShow(true);
 
 	return (
-		<>
-			<Modal show={show} onHide={handleClose}>
-        <Modal.Header closeButton>
-          <Modal.Title>Add a new channel</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <Form>
-            <Form.Group className="mb-3">
-              <Form.Label>Please type the channel name below</Form.Label>
-              <Form.Control
-                type="text"
-                onChange={(e) => setChannelAdd(e.target.value)}
-                placeholder="#"
-                required
-              />
-            </Form.Group>
-          </Form>
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={handleClose}>
-            Cancel
-          </Button>
-          <Button variant="danger" onClick={handleClose}>
-            Add
-          </Button>
-        </Modal.Footer>
-      </Modal>
 
 			<ul className="navbar-nav justify-content-end flex-grow-1 pe-3">
 				<li className="nav-item">
-					<a href="#" onClick={handleShow}>Add a Cohort</a>
+					{/* <a href="#" onClick={handleShow}>Add a Cohort</a> */}
 				</li>
 				<li className="nav-item">Logout</li>
 			</ul>
-		</>
 	);
 };
