+++ client/src/Components/StudentView/Students.css
@@ -14,6 +14,11 @@
     padding: .5em;
 }
 
+.search-sort-buttons{
+    display: flex;
+    justify-content: space-between;
+}
+
 table {
     border-collapse: separate;
     border-spacing:0 11px;
